requirements.txt
Flask==3.1.0
Flask_SocketIO==5.5.1
openpyxl==3.1.5
PyQt5==5.15.11
PyQt5_sip==12.17.0
pyqt_toast_notification==1.3.3
PySide6==6.9.0
python-socketio==5.13.0
Requests==2.32.3
